+++ packages/react/src/patterns/OneDataCollection/visualizations/collection/EditableTable/components/cells/status/NonEditableCell.tsx
@@ -1,7 +1,7 @@
-import { renderProperty } from "@/patterns/OneDataCollection/property-render"
 import { RecordType } from "@/hooks/datasource/types/records.typings"
 import { useI18n } from "@/lib/providers/i18n/i18n-provider"
 import { cn } from "@/lib/utils"
+import { renderProperty } from "@/patterns/OneDataCollection/property-render"
 
 import { EditableCellProps } from ".."
 import { BaseCell } from "../BaseCell"
@@ -14,12 +14,11 @@ export function NonEditableCell<R extends RecordType>({
   const i18n = useI18n()
 
   return (
-    <BaseCell showRightBorder={!isLastColumn}>
+    <BaseCell showRightBorder={!isLastColumn} borderOnHover={false}>
       <div
         className={cn(
           "flex w-full min-w-0",
-          "cursor-text items-center",
-          "cursor-pointer w-full",
+          "cursor-pointer items-center px-3",
           editableColumn.align === "right" && "justify-end"
         )}
       >